+++ lib/src/reducers/default-entity-collection-reducer-methods.ts
@@ -167,7 +167,7 @@ export class DefaultEntityCollectionReducerMethods<T> {
   ): EntityCollection<T> {
     return {
       ...this.adapter.addAll(action.payload, collection),
-      loaded: true, // Only QUERY_ALL_SUCCESS sets loaded to true
+      loaded: true, // only QUERY_ALL_SUCCESS and ADD_ALL sets loaded to true
       loading: false,
       originalValues: {}
     };
@@ -418,7 +418,11 @@ export class DefaultEntityCollectionReducerMethods<T> {
     action: EntityAction<T[]>
   ): EntityCollection<T> {
     this.guard.mustBeEntities(action);
-    return this.adapter.addAll(action.payload, collection);
+    return {
+      ...this.adapter.addAll(action.payload, collection),
+      loaded: true, // only QUERY_ALL_SUCCESS and ADD_ALL sets loaded to true
+      originalValues: {}
+    };
   }
 
   protected addMany(